+++ src/pages/analytics/report/Report.tsx
@@ -1,16 +1,8 @@
-import { useState, useEffect, useMemo } from 'react';
+import { useMemo } from 'react';
 import { useLocation } from 'react-router-dom';
-import {
-  Radar,
-  RadarChart,
-  PolarGrid,
-  PolarAngleAxis,
-  PolarRadiusAxis,
-  Legend,
-  ResponsiveContainer,
-} from 'recharts';
 import useUserStore from '@/store/user';
 import downloadPDF from '@/utils/downloadPDF';
+import CustomRadarChart from '@/components/radarChart/RadarChart';
 import Button from '@/components/button/Button';
 import styles from './Report.module.css';
 
@@ -58,8 +50,6 @@ const formatBulletList = (text: string): string => {
 };
 
 const Report = () => {
-  const [animate, setAnimate] = useState(true);
-
   const location = useLocation();
   const { jobField, scoreResult = [], AIAnalysis = '' } = location.state || {};
 
@@ -97,12 +87,6 @@ const Report = () => {
     [AIAnalysis],
   );
 
-  // 컴포넌트 마운트 후 애니메이션은 한 번만 실행
-  useEffect(() => {
-    const timer = setTimeout(() => setAnimate(false), 2000); // 2초 후 애니메이션 끔
-    return () => clearTimeout(timer);
-  }, []);
-
   return (
     <main className={styles.reportContainer}>
       <div id="pdf-section" className={styles.report}>
@@ -131,48 +115,11 @@ const Report = () => {
             </div>
           </div>
           <div className={styles.chart}>
-            <ResponsiveContainer width="100%" aspect={1}>
-              <RadarChart cx="50%" cy="50%" outerRadius="60%" data={data}>
-                <PolarGrid stroke="#d8d8d8" />
-                <PolarAngleAxis
-                  dataKey="subject"
-                  tick={{ fontSize: 14, fill: '#757575' }}
-                />
-                <PolarRadiusAxis domain={[0, 20]} style={{ fontSize: 10 }} />
-                <Radar
-                  name="내 역량의 평균"
-                  dataKey="averageScore"
-                  stroke="#757575"
-                  strokeDasharray={4}
-                  fillOpacity={0}
-                  dot={{
-                    fill: '#fff',
-                    fillOpacity: 1,
-                    stroke: '#757575',
-                    strokeDasharray: 0,
-                    r: 2,
-                  }}
-                  activeDot={false}
-                  isAnimationActive={animate}
-                />
-                <Radar
-                  name="값"
-                  dataKey="score"
-                  stroke="#5f812e"
-                  fill="#5f812e"
-                  fillOpacity={0.4}
-                  dot={{
-                    r: 2,
-                    fill: '#5f812e',
-                    stroke: '#5f812e',
-                    strokeWidth: 1,
-                  }}
-                  activeDot={false}
-                  isAnimationActive={animate}
-                />
-                <Legend wrapperStyle={{ fontSize: 14 }} />
-              </RadarChart>
-            </ResponsiveContainer>
+            <CustomRadarChart
+              data={data}
+              averageLabel="내 역량의 평균"
+              scoreLabel="값"
+            />
           </div>
         </section>
         <section className={styles.result}>
