+++ src/styles/global.css
@@ -40,3 +40,7 @@ body {
   padding: 0;
   position: absolute;
 }
+
+.recharts-wrapper * {
+  outline: none;
+}
